+++ lingyia_kit/tools/__init__.py
@@ -1,11 +1,13 @@
 from .fs import list_dir_tool, read_file_tool, write_file_tool
 from .http import fetch_url_tool
 from .shell import run_shell_tool
+from .sub_agent import sub_agent_tool
 
 __all__ = [
     "fetch_url_tool",
     "list_dir_tool",
     "read_file_tool",
     "run_shell_tool",
+    "sub_agent_tool",
     "write_file_tool",
 ]